chore: move emotes to enum file
