--- src/assets/interfaces/Emotes.ts
@@ -1,7 +1,0 @@
-export enum Emotes {
-  question = 'Grin',
-  confession = 'Think',
-  flirt = 'Naomato',
-  compliment = 'Blush',
-  never = 'Huh',
-}
